ADD: enable parameter NUM_REPLICAS and add function to get it from conf file

+++ src/zpack.pb-c.h
@@ -40,7 +40,7 @@ struct  _ZPack
   protobuf_c_boolean has_client_port;
   int32_t client_port;
   protobuf_c_boolean has_seq_num;
-  int32_t seq_num;
+  int64_t seq_num;
 };
 #define ZPACK__INIT \
  { PROTOBUF_C_MESSAGE_INIT (&zpack__descriptor) \
